// types/generated/graphql.tsx
export type GetUsersQuery = {
  __typename?: 'query_root'
  users: Array<{ __typename?: 'users'; id: any; name: string; created_at: any }>
}
// ...
export type CreateUserMutation = {
  __typename?: 'mutation_root'
  insert_users_one?:
    | { __typename?: 'users'; id: any; name: string; created_at: any }
    | null
    | undefined
}

export type DeleteUserMutationVariables = Exact<{
  id: Scalars['uuid']
}>

export type DeleteUserMutation = {
  __typename?: 'mutation_root'
  delete_users_by_pk?:
    | { __typename?: 'users'; id: any; name: string; created_at: any }
    | null
    | undefined
}

export type UpdateUserMutationVariables = Exact<{
  id: Scalars['uuid']
  name: Scalars['String']
}>

export type UpdateUserMutation = {
  __typename?: 'mutation_root'
  update_users_by_pk?:
    | { __typename?: 'users'; id: any; name: string; created_at: any }
    | null
    | undefined
}